requirements.txt
asgiref==3.8.1
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
cryptography==44.0.3
dj-database-url==2.3.0
Django==5.2
django-cors-headers==4.7.0
django-extensions==4.1
django-rest-passwordreset==1.5.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
gunicorn==23.0.0
idna==3.10
MarkupSafe==3.0.2
packaging==25.0
pillow==11.2.1
pycparser==2.22
PyJWT==2.9.0
pyOpenSSL==25.0.0
python-decouple==3.8
python-dotenv==1.1.0
requests==2.32.3
sqlparse==0.5.3
stripe==12.1.0
typing_extensions==4.13.2
urllib3==2.4.0
Werkzeug==3.1.3
whitenoise==6.9.0
